scripts/requirements.txt
pandas==2.0.0
tqdm==4.65.0
psycopg2==2.9.6
sqlalchemy==2.0.10
python-dotenv==1.0.0